Inicio desenvolvimento envio automatico de boletos

+++ Java/src/br/com/sankhya/truss/simprod/SimProdAction.java
@@ -3,6 +3,7 @@ package br.com.sankhya.truss.simprod;
 import java.math.BigDecimal;
 import java.math.MathContext;
 import java.sql.ResultSet;
+
 import br.com.sankhya.extensions.actionbutton.AcaoRotinaJava;
 import br.com.sankhya.extensions.actionbutton.ContextoAcao;
 import br.com.sankhya.extensions.actionbutton.Registro;
@@ -28,67 +29,67 @@ public class SimProdAction implements AcaoRotinaJava {
 		SessionHandle hnd = null;
 		JdbcWrapper jdbc = null;
 		DynamicVO proVO = null;
-		
+
 		try {
 			hnd = JapeSession.open();
 			EntityFacade dwfFacade = EntityFacadeFactory.getDWFFacade();
 			jdbc = dwfFacade.getJdbcWrapper();
 			jdbc.openSession();
-			
+
 			Registro[] linhas = ctx.getLinhas();
-			
-			
-			
+
+
+
 			for(Registro linha: linhas) {
-				
+
 				BigDecimal lotepad = (BigDecimal) linha.getCampo("TAMLOTEPAD");
-						
+
 				NativeSql q = new NativeSql(jdbc);
 				ResultSet r = null;
-				
-				r = q.executeQuery("SELECT VW.PRODUTO, VW.CODPRODMP, SUM(VW.QTDMISTURA) AS QTDMISTURA, VW.CODVOL FROM AD_ARVOREPRODUTO VW JOIN TPRPRC PRC ON PRC.CODPRC = VW.CODPRC AND PRC.VERSAO = VW.VERSAO AND PRC.PADRAO = 'S' WHERE PRODUTO = " + (BigDecimal) linha.getCampo("CODPRODPA") + " GROUP BY VW.PRODUTO, VW.CODPRODMP, VW.CODVOL");
-				
-				
+
+				r = q.executeQuery("SELECT VW.PRODUTO, VW.CODPRODMP, SUM(VW.QTDMISTURA) AS QTDMISTURA, VW.CODVOL FROM AD_ARVOREPRODUTO VW JOIN TPRPRC PRC ON PRC.CODPRC = VW.CODPRC AND PRC.VERSAO = VW.VERSAO AND PRC.PADRAO = 'S' WHERE PRODUTO = " + linha.getCampo("CODPRODPA") + " GROUP BY VW.PRODUTO, VW.CODPRODMP, VW.CODVOL");
+
+
 				while(r.next()){
 					NativeSql query = new NativeSql(jdbc);
 					ResultSet rs = null;
-					
+
 					proVO = proDAO.findByPK(r.getBigDecimal("CODPRODMP"));
-					
+
 					NativeSql q2 = new NativeSql(jdbc);
 					ResultSet r2 = null;
 
-					r2 = q2.executeQuery("SELECT SUM(ESTOQUE - RESERVADO) AS ESTOQUE\r\n" + 
-							"					FROM TGFEST\r\n" + 
-							"					WHERE CODEMP = 6\r\n" + 
-							"					AND CODLOCAL = 11100" + 
-							"					AND CODPARC = 0" + 
+					r2 = q2.executeQuery("SELECT SUM(ESTOQUE - RESERVADO) AS ESTOQUE\r\n" +
+							"					FROM TGFEST\r\n" +
+							"					WHERE CODEMP = 6\r\n" +
+							"					AND CODLOCAL = 11100" +
+							"					AND CODPARC = 0" +
 							"					AND CODPROD = " + r.getBigDecimal("CODPRODMP"));
-					
+
 					BigDecimal estoque = BigDecimal.ZERO;
-					
+
 					if(r2.next()) {
 						estoque = r2.getBigDecimal("ESTOQUE");
 					}
-					
+
 					BigDecimal estmax = proVO.asBigDecimal("ESTMAX");
 					BigDecimal estmin = proVO.asBigDecimal("ESTMIN");
 					BigDecimal qtdpedido = BigDecimal.ZERO;
-					
-					rs = query.executeQuery("SELECT NVL(SUM(ITE.QTDNEG),0) AS QUANTIDADE\r\n" + 
-							"FROM TGFITE ITE\r\n" + 
-							"JOIN TGFCAB CAB ON CAB.NUNOTA = ITE.NUNOTA\r\n" + 
-							"WHERE CAB.TIPMOV = 'O'\r\n" + 
-							"AND CAB.STATUSNOTA = 'L'\r\n" + 
-							"AND ITE.PENDENTE = 'S'\r\n" + 
+
+					rs = query.executeQuery("SELECT NVL(SUM(ITE.QTDNEG),0) AS QUANTIDADE\r\n" +
+							"FROM TGFITE ITE\r\n" +
+							"JOIN TGFCAB CAB ON CAB.NUNOTA = ITE.NUNOTA\r\n" +
+							"WHERE CAB.TIPMOV = 'O'\r\n" +
+							"AND CAB.STATUSNOTA = 'L'\r\n" +
+							"AND ITE.PENDENTE = 'S'\r\n" +
 							"AND ITE.CODPROD = " + r.getBigDecimal("CODPRODMP"));
-					
+
 					if(rs.next()) {
 						qtdpedido = rs.getBigDecimal("QUANTIDADE");
 					}
-					
+
 					simiteDAO.create()
-					.set("NUSIM", (BigDecimal) linha.getCampo("NUSIM"))
+					.set("NUSIM", linha.getCampo("NUSIM"))
 					.set("CODPROD", r.getBigDecimal("CODPRODMP"))
 					.set("QTDMISTURA", r.getBigDecimal("QTDMISTURA").multiply(lotepad, new MathContext(6)))
 					.set("CODVOL", r.getString("CODVOL"))
@@ -97,8 +98,8 @@ public class SimProdAction implements AcaoRotinaJava {
 					.set("ESTOQUE", estoque)
 					.set("QTDPEDCOMPRA", qtdpedido)
 					.save();
-					
-					
+
+
 				}
 			}
 			ctx.setMensagemRetorno("Simulação Concluída.");
